\name{adjust_times}
\alias{adjust_times}
\title{Adjust UTC times relative to current time zone}
\usage{
  adjust_times(x, hours)
}
\arguments{
  \item{x}{POSIXct dates to be converted}

  \item{hours}{a positive or negative integer indicating
  how many hours the times should be adjusted}
}
\description{
  Adjust UTC times relative to current time zone
}

