\name{punchcard}
\alias{punchcard}
\title{Punchcard plot of usage by time and time}
\usage{
  punchcard(x, all.day = TRUE)
}
\arguments{
  \item{x}{alfred object}

  \item{all.day}{logical, should plot span 24 hours
  regardless of activity?}
}
\description{
  Punchcard plot of usage by time and time
}

